Upload: Move CLI tests to test/upload/cli
