+++ test/upload/cli/test_select.py
@@ -7,7 +7,7 @@ from argparse import Namespace
 import six
 import tweak
 
-from .. import CapturingIO, reset_tweak_changes
+from ... import CapturingIO, reset_tweak_changes
 
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
